Change parse calculate name and add user logs
